fix: brc-20 tokens breaking ledger
