+++ src/app/components/crypto-assets/bitcoin/brc20-token-asset-list/components/brc20-token-asset-item.layout.tsx
@@ -4,14 +4,13 @@ import { forwardRefWithAs } from '@stacks/ui-core';
 import type { Money } from '@shared/models/money.model';
 
 import { formatBalance } from '@app/common/format-balance';
+import { AssetCaption } from '@app/components/crypto-assets/components/asset-caption';
 import { Brc20TokenIcon } from '@app/components/icons/brc20-token-icon';
 import { usePressable } from '@app/components/item-hover';
 import { Flag } from '@app/components/layout/flag';
 import { SpaceBetween } from '@app/components/layout/space-between';
 import { Tooltip } from '@app/components/tooltip';
 import { Text } from '@app/components/typography';
-
-import { AssetCaption } from '../../components/asset-caption';
 
 interface Brc20TokenAssetItemLayoutProps extends BoxProps {
   balance: Money;
