+++ src/app/features/balances-list/components/bitcoin-fungible-tokens-asset-list.tsx
@@ -1,6 +1,6 @@
 import { Stack } from '@stacks/ui';
 
-import { Brc20TokenAssetItem } from '@app/components/crypto-assets/bitcoin/brc20-token-asset/brc20-token-asset-item';
+import { Brc20TokenAssetItem } from '@app/components/crypto-assets/bitcoin/brc20-token-asset-list/components/brc20-token-asset-item';
 import { Brc20Token } from '@app/query/bitcoin/ordinals/brc20/brc20-tokens.query';
 
 interface BitcoinFungibleTokenAssetListProps {
